+++ project/Versions.scala
@@ -2,11 +2,11 @@ object Versions {
 
   val Scala_2_11 = "2.11.12"
 
-  val Scala_2_13 = "2.13.4"
+  val Scala_2_13 = "2.13.5"
 
-  val Scala_2_12 = "2.12.12"
+  val Scala_2_12 = "2.12.15"
 
   val Scala_3 = "3.0.0"
 
-  val ScalaJsDom = "1.1.0"
+  val ScalaJsDom = "2.0.0"
 }